+++ angel_orm/example/main.dart
@@ -34,7 +34,7 @@ class _FakeExecutor extends QueryExecutor {
   }
 
   @override
-  Future<T> transaction<T>(FutureOr<T> Function() f) {
+  Future<T> transaction<T>(FutureOr<T> Function(QueryExecutor) f) {
     throw UnsupportedError('Transactions are not supported.');
   }
 }