+++ angel_orm/lib/src/query_executor.dart
@@ -11,6 +11,13 @@ abstract class QueryExecutor {
       String tableName, String query, Map<String, dynamic> substitutionValues,
       [List<String> returningFields]);
 
-  /// Begins a database transaction.
-  Future<T> transaction<T>(FutureOr<T> f());
+  /// Enters a database transaction, performing the actions within,
+  /// and returning the results of [f].
+  ///
+  /// If [f] fails, the transaction will be rolled back, and the
+  /// responsible exception will be re-thrown.
+  /// 
+  /// Whether nested transactions are supported depends on the
+  /// underlying driver.
+  Future<T> transaction<T>(FutureOr<T> Function(QueryExecutor) f);
 }