+++ angel_orm/CHANGELOG.md
@@ -1,6 +1,8 @@
 # 2.1.0
 * Split the formerly 600+ line `src/query.dart` up into
 separate files.
+* **BREAKING**: Add a required `QueryExecutor` argument to `transaction`
+callbacks.
 
 # 2.0.1
 * Apply `package:pedantic` fixes.